Created ButtomMenu component and moved all of the List component's button's styles and functionality into it closes #62
