+++ src/components/ButtonMenu/ButtonMenu.style.js
@@ -1,0 +1,6 @@
+import { StyleSheet } from 'react-native'
+
+export default StyleSheet.create({
+  view: {},
+  text: {},
+})
